butterfly/submission/requirements.txt
open-clip-torch==2.24.0
torch==2.3.0
torchvision==0.18.0